This is the actual code, the previous commit was nonsense.  This version implements a lock free structure, but it is an imperfect LRU.  That means that if two threads add a new element in the queue at the same time, the two new entries might have the same age. So the replacement policy is not really "Least Recently Used" but rather "Not Most Recently Used" as it will still evict older elements before evicting newer.
Another side-effect of the lock free mechanism is that it does _not_ guarantee
fairness. A thread could suffer starvation. However, this is not expected to have
a significant enough probability to cause a real effect on performance.

The positive result of the lock free approach is the speed.  The "driver" test shows
a latency of 1.3 usec per operation on zoot, with 6 threads sharing an L3 and
having private L2 caches.