Fix the filling up of assignments array in the case of the startup tasks.
An inline function might be used to define a variable before all variables are defined. Thus, the assignments array must be filled as we go and define the variables, not in one go when all variables are defined.

Similar errors might have been done at other places in the code. Debugging in progress.