Add a Priority Based Local Flat Queue (PBQ) scheduler.   This scheduler acts exactly as LFQ, except that the push operation is done by priority: tasks with a high priority will expel tasks with a low priority to the system queue.
Nota Bene: this asumes that the task queue given to the scheduler is sorted *in reverse order*.

The fact that the queue *must be sorted in reverse order* (bigger priority number first) to get the expected behavior in the push is probably the reason why there is/was confusion in the code on what is a high priority.

A high priority is a priority with a high number. When building the list of tasks to be scheduled, tasks with a high priority must come first.