The MCA parameters from Open MPI, reshaped to fit the DAGuE framework. The parameters can be read from a file, retrieved from environment variable (or the registry on Windows), or set by the command line. This provide a consistent and easy to use framework to deal with all these parameters.
BEWARE: The building of the database is not thread safe, so this
should be done during the initialization, at a moment when there is
only one active thread. The reading is thread-safe. The destruction
is not, but as the entire database is destroyed only once, this
doesn't really matter.

This is imported from Open MPI.