We're back in shape. Some of the object are now derived from the dague_object_t base obejct (which is a new concept different from the old dague_object_t).
The old dague_object_t type has been renamed to dague_handle_t
which is a better fit (from my pespective at least). It does
not point to an object per see, but to a externally visible
DAGuE function. Moreover, the external layers has no knowledge
about the internals, so it is indeed mostly a handle.

The current status is that everything compiles and runs on shared
memory. The GPU and distributed will be the next step. I converted
almost all old concepts to the new data centric approach, with a
single exception: the RBT, more exactly the data_of. There an offset
is used in a strange way, this made me skeptical about my changes,
so they are not committed here. I'll ping Anthony to see how to fix
this.